# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FocusGuard AI API - Focus session management with Google Calendar and Gemini AI
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FocusSession {
  session_id: string;
  event_title: string;
  start_time: string;
  end_time: string;
  focus_intensity: string;
  blocked_sites: string[];
  emergency_sites: string[];
  is_active: boolean;
  /** @nullable */
  ai_reason?: string | null;
  completed?: boolean;
  completion_verified?: boolean;
  user_email?: string;
  /** @nullable */
  event_id?: string | null;
  override_count?: number;
}

export interface SyncResult {
  sessions_created: number;
  sessions_updated: number;
  message: string;
  events_analyzed?: number;
}

export interface OverrideInput {
  user_email: string;
  reason: string;
}

export interface OverrideResult {
  success: boolean;
  message: string;
}

export interface CompletionInput {
  user_email: string;
  description?: string;
  image_base64?: string;
}

export interface CompletionResult {
  verified: boolean;
  message: string;
  /** @nullable */
  ai_analysis?: string | null;
  /** @nullable */
  confidence?: number | null;
}

export interface UserSettings {
  user_email: string;
  blocked_sites: string[];
  emergency_sites: string[];
  trigger_intensity: string;
  advance_minutes: number;
  calendar_connected?: boolean;
  /** @nullable */
  google_access_token?: string | null;
}

export interface UserSettingsInput {
  blocked_sites?: string[];
  emergency_sites?: string[];
  trigger_intensity?: string;
  advance_minutes?: number;
  /** @nullable */
  google_access_token?: string | null;
}

export interface DashboardSummary {
  active_session: FocusSession | null;
  upcoming_sessions: FocusSession[];
  total_sessions_today: number;
  focus_minutes_today: number;
  top_blocked_sites: string[];
}

export interface CalendarEvent {
  event_id: string;
  title: string;
  start_time: string;
  end_time: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  ai_focus_score?: number | null;
  /** @nullable */
  suggested_intensity?: string | null;
}