# lib/api-zod/src/generated/types/calendarEvent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FocusGuard AI API - Focus session management with Google Calendar and Gemini AI
 * OpenAPI spec version: 0.1.0
 */

export interface CalendarEvent {
  event_id: string;
  title: string;
  start_time: string;
  end_time: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  ai_focus_score?: number | null;
  /** @nullable */
  suggested_intensity?: string | null;
}