# lib/api-zod/src/generated/types/completionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FocusGuard AI API - Focus session management with Google Calendar and Gemini AI
 * OpenAPI spec version: 0.1.0
 */

export interface CompletionInput {
  user_email: string;
  description?: string;
  image_base64?: string;
}