# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FocusGuard AI API - Focus session management with Google Calendar and Gemini AI
 * OpenAPI spec version: 0.1.0
 */
import type { FocusSession } from './focusSession';

export interface DashboardSummary {
  active_session: FocusSession | null;
  upcoming_sessions: FocusSession[];
  total_sessions_today: number;
  focus_minutes_today: number;
  top_blocked_sites: string[];
}