# lib/api-zod/src/generated/types/focusSession.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FocusGuard AI API - Focus session management with Google Calendar and Gemini AI
 * OpenAPI spec version: 0.1.0
 */

export interface FocusSession {
  session_id: string;
  event_title: string;
  start_time: string;
  end_time: string;
  focus_intensity: string;
  blocked_sites: string[];
  emergency_sites: string[];
  is_active: boolean;
  /** @nullable */
  ai_reason?: string | null;
  completed?: boolean;
  completion_verified?: boolean;
  user_email?: string;
  /** @nullable */
  event_id?: string | null;
  override_count?: number;
}