# lib/api-zod/src/generated/types/overrideInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FocusGuard AI API - Focus session management with Google Calendar and Gemini AI
 * OpenAPI spec version: 0.1.0
 */

export interface OverrideInput {
  user_email: string;
  reason: string;
}