# lib/api-zod/src/generated/types/syncResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FocusGuard AI API - Focus session management with Google Calendar and Gemini AI
 * OpenAPI spec version: 0.1.0
 */

export interface SyncResult {
  sessions_created: number;
  sessions_updated: number;
  message: string;
  events_analyzed?: number;
}