# lib/api-zod/src/generated/types/userSettings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FocusGuard AI API - Focus session management with Google Calendar and Gemini AI
 * OpenAPI spec version: 0.1.0
 */

export interface UserSettings {
  user_email: string;
  blocked_sites: string[];
  emergency_sites: string[];
  trigger_intensity: string;
  advance_minutes: number;
  calendar_connected?: boolean;
  /** @nullable */
  google_access_token?: string | null;
}